Create componentDidMount example
There is a mock API call. 
The best practice with React is to place API calls or any calls to server in the lifecycle method componentDidMount(). 
This method is called after a component is mounted to the DOM. 
Any calls to setState() here will trigger a re-rendering of your component. 
When you call an API in this method, and set your state with the data that the API returns, it will automatically trigger an update once you receive the data.